Modify vakidation

See merge request comma/gostaresh!40

+++ database/migrations/2022_06_22_172706_create_gostaresh_amount_of_facilities_for_research_achievements_table.php
@@ -24,7 +24,7 @@ return new class extends Migration
             $table->foreignId('rural_district_id')->nullable();
 
             $table->string('unit')->nullable();
-            $table->unsignedBigInteger('amount')->nullable();
+            $table->float('amount')->nullable();
 
             $table->integer('year')->unsigned()->nullable();
             $table->tinyInteger('month')->unsigned()->nullable();
