+++ database/migrations/2022_06_25_073408_create_gostaresh_number_of_students_status_analyses_table.php
@@ -30,9 +30,9 @@ return new class extends Migration
             $table->foreignId('minor_id')->nullable();
 
             $table->tinyInteger('university_type')->nullable();
-            $table->unsignedBigInteger('number_of_students')->nullable();
-            $table->tinyInteger('department_of_education')->nullable(); 
-            
+            $table->float('number_of_students')->nullable();
+            $table->tinyInteger('department_of_education')->nullable();
+
             $table->integer('year')->unsigned()->nullable();
             $table->tinyInteger('month')->unsigned()->nullable();
             $table->softDeletes();
