+++ database/migrations/2022_06_25_200212_create_gostaresh_status_analysis_of_the_number_of_courses_table.php
@@ -24,11 +24,11 @@ return new class extends Migration
             $table->foreignId('rural_district_id')->nullable();
 
             $table->string('unit')->nullable();
-            $table->unsignedBigInteger('total_number_of_courses')->comment('تعداد کل دوره های تحصیلی')->nullable();
-            $table->unsignedBigInteger('number_of_international_Persian_language_courses_in_person')->comment('تعداد دوره های تحصیلی بین المللی زبان فارسی  حضوری')->nullable();
-            $table->unsignedBigInteger('number_of_international_virtual_Persian_language_courses')->comment('داد دوره های تحصیلی بین المللی برگزار شده به زبان فارسی به صورت مجازی')->nullable();
-            $table->unsignedBigInteger('number_of_international_courses_in_the_target_language_in_person')->comment('تعداد دوره های تحصیلی بین المللی برگزار شده به زبان های کشورهای هدف به صورت حضوری')->nullable();
-            $table->unsignedBigInteger('number_of_international_courses_in_the_target_language_virtually')->comment('تعداد دوره های تحصیلی بین المللی برگزار شده به زبان های کشورهای هدف به صورت مجازی')->nullable();
+            $table->float('total_number_of_courses')->comment('تعداد کل دوره های تحصیلی')->nullable();
+            $table->float('number_of_international_Persian_language_courses_in_person')->comment('تعداد دوره های تحصیلی بین المللی زبان فارسی  حضوری')->nullable();
+            $table->float('number_of_international_virtual_Persian_language_courses')->comment('داد دوره های تحصیلی بین المللی برگزار شده به زبان فارسی به صورت مجازی')->nullable();
+            $table->float('number_of_international_courses_in_the_target_language_in_person')->comment('تعداد دوره های تحصیلی بین المللی برگزار شده به زبان های کشورهای هدف به صورت حضوری')->nullable();
+            $table->float('number_of_international_courses_in_the_target_language_virtually')->comment('تعداد دوره های تحصیلی بین المللی برگزار شده به زبان های کشورهای هدف به صورت مجازی')->nullable();
 
             $table->integer('year')->unsigned()->nullable();
             $table->tinyInteger('month')->unsigned()->nullable();
